SECURE AUTHENTICATION SYSTEM/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31 # Added explicitly for clarity, usually pulled by Flask-SQLAlchemy